Added RocksDb data source option

+++ rskj-core/src/main/java/org/ethereum/datasource/RocksDbDataSource.java
@@ -1,6 +1,6 @@
 /*
  * This file is part of RskJ
- * Copyright (C) 2017 RSK Labs Ltd.
+ * Copyright (C) 2021 RSK Labs Ltd.
  * (derived from ethereumJ library, Copyright (c) 2016 <ether.camp>)
  *
  * This program is free software: you can redistribute it and/or modify
@@ -24,7 +24,7 @@ import co.rsk.metrics.profilers.ProfilerFactory;
 import co.rsk.panic.PanicProcessor;
 import org.ethereum.db.ByteArrayWrapper;
 import org.ethereum.util.ByteUtil;
-import org.iq80.leveldb.*;
+import org.rocksdb.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -40,17 +40,20 @@ import java.util.concurrent.locks.ReadWriteLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 import static java.lang.System.getProperty;
-import static org.fusesource.leveldbjni.JniDBFactory.factory;
 
-public class LevelDbDataSource implements KeyValueDataSource {
+public class RocksDbDataSource implements KeyValueDataSource {
 
     private static final Logger logger = LoggerFactory.getLogger("db");
     private static final Profiler profiler = ProfilerFactory.getInstance();
     private static final PanicProcessor panicProcessor = new PanicProcessor();
 
+    static {
+        RocksDB.loadLibrary();
+    }
+
     private final String databaseDir;
     private final String name;
-    private DB db;
+    private RocksDB db;
     private boolean alive;
 
     // The native LevelDB insert/update/delete are normally thread-safe
@@ -59,12 +62,13 @@ public class LevelDbDataSource implements KeyValueDataSource {
     // The leveldbJNI lib has a protection over accessing closed DB but it is not synchronized
     // This ReadWriteLock still permits concurrent execution of insert/delete/update operations
     // however blocks them on init/close/delete operations
+    // TODO: check if this lock is needed
     private final ReadWriteLock resetDbLock = new ReentrantReadWriteLock();
 
-    public LevelDbDataSource(String name, String databaseDir) {
+    public RocksDbDataSource(String name, String databaseDir) {
         this.databaseDir = databaseDir;
         this.name = name;
-        logger.debug("New LevelDbDataSource: {}", name);
+        logger.debug("New RocksDbDataSource: {}", name);
     }
 
     @Override
@@ -72,7 +76,7 @@ public class LevelDbDataSource implements KeyValueDataSource {
         resetDbLock.writeLock().lock();
         Metric metric = profiler.start(Profiler.PROFILING_TYPE.LEVEL_DB_INIT);
         try {
-            logger.debug("~> LevelDbDataSource.init(): {}", name);
+            logger.debug("~> RocksDbDataSource.init(): {}", name);
 
             if (isAlive()) {
                 return;
@@ -81,13 +85,13 @@ public class LevelDbDataSource implements KeyValueDataSource {
             Objects.requireNonNull(name, "no name set to the db");
 
             Options options = new Options();
-            options.createIfMissing(true);
-            options.compressionType(CompressionType.NONE);
-            options.blockSize(10 * 1024 * 1024);
-            options.writeBufferSize(10 * 1024 * 1024);
-            options.cacheSize(0);
-            options.paranoidChecks(true);
-            options.verifyChecksums(true);
+            options.setCreateIfMissing(true);
+            options.setCompressionType(CompressionType.NO_COMPRESSION);
+//            options.setArenaBlockSize(10 * 1024 * 1024); TODO: check if this is needed
+            options.setWriteBufferSize(10 * 1024 * 1024);
+//            options.cacheSize(0); TODO: check if this is needed
+            options.setParanoidChecks(true);
+//            options.verifyChecksums(true); TODO: check if this is needed
 
             try {
 
@@ -97,22 +101,22 @@ public class LevelDbDataSource implements KeyValueDataSource {
                 Files.createDirectories(dbPath.getParent());
 
                 logger.debug("Initializing new or existing database: '{}'", name);
-                db = factory.open(dbPath.toFile(), options);
+                db = RocksDB.open(options, dbPath.toString());
 
                 alive = true;
-            } catch (IOException ioe) {
-                logger.error(ioe.getMessage(), ioe);
-                panicProcessor.panic("leveldb", ioe.getMessage());
-                throw new RuntimeException("Can't initialize database");
+            } catch (IOException | RocksDBException e) {
+                logger.error(e.getMessage(), e);
+                panicProcessor.panic("rocksdb", e.getMessage());
+                throw new DataSourceException("Can't initialize database", e);
             }
-            logger.debug("<~ LevelDbDataSource.init(): {}", name);
+            logger.debug("<~ RocksDbDataSource.init(): {}", name);
         } finally {
             profiler.stop(metric);
             resetDbLock.writeLock().unlock();
         }
     }
 
-    public static Path getPathForName(String name, String databaseDir) {
+    private static Path getPathForName(String name, String databaseDir) {
         if (Paths.get(databaseDir).isAbsolute()) {
             return Paths.get(databaseDir, name);
         } else {
@@ -142,29 +146,33 @@ public class LevelDbDataSource implements KeyValueDataSource {
         resetDbLock.readLock().lock();
         try {
             if (logger.isTraceEnabled()) {
-                logger.trace("~> LevelDbDataSource.get(): {}, key: {}", name,  ByteUtil.toHexString(key));
+                logger.trace("~> RocksDbDataSource.get(): {}, key: {}", name,  ByteUtil.toHexString(key));
+            }
+
+            if (!alive) {
+                throw new DataSourceException("Db is not alive");
             }
 
             try {
                 byte[] ret = db.get(key);
                 if (logger.isTraceEnabled()) {
-                    logger.trace("<~ LevelDbDataSource.get(): {}, key: {}, return length: {}", name, ByteUtil.toHexString(key), (ret == null ? "null" : ret.length));
+                    logger.trace("<~ RocksDbDataSource.get(): {}, key: {}, return length: {}", name, ByteUtil.toHexString(key), (ret == null ? "null" : ret.length));
                 }
 
                 return ret;
-            } catch (DBException e) {
+            } catch (RocksDBException e) {
                 logger.error("Exception. Retrying again...", e);
                 try {
                     byte[] ret = db.get(key);
                     if (logger.isTraceEnabled()) {
-                        logger.trace("<~ LevelDbDataSource.get(): {}, key: {}, return length: {}", name, ByteUtil.toHexString(key), (ret == null ? "null" : ret.length));
+                        logger.trace("<~ RocksDbDataSource.get(): {}, key: {}, return length: {}", name, ByteUtil.toHexString(key), (ret == null ? "null" : ret.length));
                     }
 
                     return ret;
-                } catch (DBException e2) {
+                } catch (RocksDBException e2) {
                     logger.error("Exception. Not retrying.", e2);
                     panicProcessor.panic("leveldb", String.format("Exception. Not retrying. %s", e2.getMessage()));
-                    throw e2;
+                    throw new DataSourceException("Get op failed", e2);
                 }
             }
         } finally {
@@ -182,15 +190,23 @@ public class LevelDbDataSource implements KeyValueDataSource {
         resetDbLock.readLock().lock();
         try {
             if (logger.isTraceEnabled()) {
-                logger.trace("~> LevelDbDataSource.put(): {}, key: {}, return length: {}", name, ByteUtil.toHexString(key), value.length);
+                logger.trace("~> RocksDbDataSource.put(): {}, key: {}, return length: {}", name, ByteUtil.toHexString(key), value.length);
+            }
+
+            if (!alive) {
+                throw new DataSourceException("Db is not alive");
             }
 
             db.put(key, value);
             if (logger.isTraceEnabled()) {
-                logger.trace("<~ LevelDbDataSource.put(): {}, key: {}, return length: {}", name, ByteUtil.toHexString(key), value.length);
+                logger.trace("<~ RocksDbDataSource.put(): {}, key: {}, return length: {}", name, ByteUtil.toHexString(key), value.length);
             }
 
             return value;
+        } catch (RocksDBException e) {
+            logger.error("Exception. Not retrying.", e);
+            panicProcessor.panic("leveldb", String.format("Exception. Not retrying. %s", e.getMessage()));
+            throw new DataSourceException("Put op failed", e);
         } finally {
             resetDbLock.readLock().unlock();
             profiler.stop(metric);
@@ -203,14 +219,22 @@ public class LevelDbDataSource implements KeyValueDataSource {
         resetDbLock.readLock().lock();
         try {
             if (logger.isTraceEnabled()) {
-                logger.trace("~> LevelDbDataSource.delete(): {}, key: {}", name, ByteUtil.toHexString(key));
+                logger.trace("~> RocksDbDataSource.delete(): {}, key: {}", name, ByteUtil.toHexString(key));
+            }
+
+            if (!alive) {
+                throw new DataSourceException("Db is not alive");
             }
 
             db.delete(key);
             if (logger.isTraceEnabled()) {
-                logger.trace("<~ LevelDbDataSource.delete(): {}, key: {}", name, ByteUtil.toHexString(key));
+                logger.trace("<~ RocksDbDataSource.delete(): {}, key: {}", name, ByteUtil.toHexString(key));
             }
 
+        } catch (RocksDBException e) {
+            logger.error("Exception. Not retrying.", e);
+            panicProcessor.panic("leveldb", String.format("Exception. Not retrying. %s", e.getMessage()));
+            throw new DataSourceException("Delete op failed", e);
         } finally {
             resetDbLock.readLock().unlock();
             profiler.stop(metric);
@@ -223,23 +247,28 @@ public class LevelDbDataSource implements KeyValueDataSource {
         resetDbLock.readLock().lock();
         try {
             if (logger.isTraceEnabled()) {
-                logger.trace("~> LevelDbDataSource.keys(): {}", name);
+                logger.trace("~> RocksDbDataSource.keys(): {}", name);
             }
 
-            try (DBIterator iterator = db.iterator()) {
+            if (!alive) {
+                throw new DataSourceException("Db is not alive");
+            }
+
+            try (RocksIterator iterator = db.newIterator()) {
                 Set<byte[]> result = new HashSet<>();
-                for (iterator.seekToFirst(); iterator.hasNext(); iterator.next()) {
-                    result.add(iterator.peekNext().getKey());
+                iterator.seekToFirst();
+                while (iterator.isValid()) {
+                    result.add(iterator.key());
                 }
                 if (logger.isTraceEnabled()) {
-                    logger.trace("<~ LevelDbDataSource.keys(): {}, {}", name, result.size());
+                    logger.trace("<~ RocksDbDataSource.keys(): {}, {}", name, result.size());
                 }
 
                 return result;
-            } catch (IOException e) {
+            } catch (RuntimeException e) {
                 logger.error("Unexpected", e);
                 panicProcessor.panic("leveldb", String.format("Unexpected %s", e.getMessage()));
-                throw new RuntimeException(e);
+                throw new DataSourceException("Cannot retrieve keys", e);
             }
         } finally {
             resetDbLock.readLock().unlock();
@@ -247,24 +276,29 @@ public class LevelDbDataSource implements KeyValueDataSource {
         }
     }
 
-    private void updateBatchInternal(Map<ByteArrayWrapper, byte[]> rows, Set<ByteArrayWrapper> deleteKeys) throws IOException {
+    private void updateBatchInternal(Map<ByteArrayWrapper, byte[]> rows, Set<ByteArrayWrapper> deleteKeys) {
         Metric metric = profiler.start(Profiler.PROFILING_TYPE.DB_WRITE);
         if (rows.containsKey(null) || rows.containsValue(null)) {
             profiler.stop(metric);
             throw new IllegalArgumentException("Cannot update null values");
         }
         // Note that this is not atomic.
-        try (WriteBatch batch = db.createWriteBatch()) {
+        try (WriteBatch batch = new WriteBatch()) {
+            WriteOptions writeOpts = new WriteOptions();
+
             for (Map.Entry<ByteArrayWrapper, byte[]> entry : rows.entrySet()) {
                 batch.put(entry.getKey().getData(), entry.getValue());
             }
             for (ByteArrayWrapper deleteKey : deleteKeys) {
                 batch.delete(deleteKey.getData());
             }
-            db.write(batch);
+            db.write(writeOpts, batch);
             profiler.stop(metric);
+        } catch (RocksDBException e) {
+            logger.error("Exception. Not retrying.", e);
+            panicProcessor.panic("leveldb", String.format("Exception. Not retrying. %s", e.getMessage()));
+            throw new DataSourceException("Update batch op failed", e);
         }
-
     }
 
     @Override
@@ -275,30 +309,34 @@ public class LevelDbDataSource implements KeyValueDataSource {
         resetDbLock.readLock().lock();
         try {
             if (logger.isTraceEnabled()) {
-                logger.trace("~> LevelDbDataSource.updateBatch(): {}, {}", name, rows.size());
+                logger.trace("~> RocksDbDataSource.updateBatch(): {}, {}", name, rows.size());
+            }
+
+            if (!alive) {
+                throw new DataSourceException("Db is not alive");
             }
 
             try {
                 updateBatchInternal(rows, deleteKeys);
                 if (logger.isTraceEnabled()) {
-                    logger.trace("<~ LevelDbDataSource.updateBatch(): {}, {}", name, rows.size());
+                    logger.trace("<~ RocksDbDataSource.updateBatch(): {}, {}", name, rows.size());
                 }
             } catch (IllegalArgumentException iae) {
                 throw iae;
-            } catch (Exception e) {
+            } catch (RuntimeException e) {
                 logger.error("Error, retrying one more time...", e);
                 // try one more time
                 try {
                     updateBatchInternal(rows, deleteKeys);
                     if (logger.isTraceEnabled()) {
-                        logger.trace("<~ LevelDbDataSource.updateBatch(): {}, {}", name, rows.size());
+                        logger.trace("<~ RocksDbDataSource.updateBatch(): {}, {}", name, rows.size());
                     }
                 } catch (IllegalArgumentException iae) {
                     throw iae;
-                } catch (Exception e1) {
+                } catch (RuntimeException e1) {
                     logger.error("Error", e);
                     panicProcessor.panic("leveldb", String.format("Error %s", e.getMessage()));
-                    throw new RuntimeException(e);
+                    throw e1;
                 }
             }
         } finally {
@@ -320,7 +358,7 @@ public class LevelDbDataSource implements KeyValueDataSource {
                 db.close();
 
                 alive = false;
-            } catch (IOException e) {
+            } catch (RuntimeException e) {
                 logger.error("Failed to find the db file on the close: {} ", name);
                 panicProcessor.panic("leveldb", String.format("Failed to find the db file on the close: %s", name));
             }
